requirements.txt
requests==2.31.0
numpy==1.26.4
python-dotenv==1.0.1
datasets==2.18.0
tweepy==4.14.0
notebook